serving/requirements.txt
fastapi
gunicorn
pandas
scikit-learn
google-cloud-storage